[MET-598, MET-599, MET-600] Apply newest code for consumer, data api and construction api

+++ api/src/main/java/org/cardanofoundation/rosetta/api/model/rest/FindUtxo.java
@@ -12,7 +12,7 @@ public class FindUtxo {
 
   private BigInteger value;
   private String txHash;
-  private Integer index;
+  private Short index;
   private String policy;
   private String name;
   private BigInteger quantity;
